Cardano.CoinSelection

Description

Provides general functions and types relating to coin selection.

The CoinSelectionAlgorithm type provides a common interface to coin selection algorithms.

The CoinSelection type represents the result of running a coin selection algorithm.

Coin

newtype Coin

A non-negative integer value that represents a number of Lovelace.

One Ada is equal to 1,000,000 Lovelace.

Constructors

Coin

Fields

unCoin :: Word64

coinIsValid :: Coin -> Bool

Coin Map

newtype CoinMap a

A mapping from unique keys to associated Coin values.

A CoinMap can be used to represent:

a UTxO set, where each key within the map refers to an unspent output from a previous transaction.

a set of inputs to a CoinSelection, where each input is an entry selected from a UTxO set by a CoinSelectionAlgorithm.

a set of outputs for a CoinSelection, where each key within the map refers to the address of a payment recipient.

A CoinMap can be constructed with the coinMapFromList function.

The total value of a CoinMap is given by the coinMapValue function.

Constructors

CoinMap

Fields

unCoinMap :: Map a Coin

data CoinMapEntry a

An entry for a CoinMap.

Constructors

CoinMapEntry

Fields

entryKey :: a

The unique key associated with this entry.

entryValue :: Coin

The coin value associated with this entry.

coinMapFromList :: Ord a => [CoinMapEntry a] -> CoinMap a

Constructs a CoinMap from a list of entries.

See CoinMapEntry.

coinMapToList :: CoinMap a -> [CoinMapEntry a]

Converts a CoinMap to a list of entries.

See CoinMapEntry.

coinMapValue :: CoinMap a -> Coin

Calculates the total coin value associated with a CoinMap.

coinMapRandomEntry :: MonadRandom m => CoinMap a -> m (Maybe (CoinMapEntry a), CoinMap a)

Selects an entry at random from a CoinMap, returning both the selected entry and the map with the entry removed.

If the given map is empty, this function returns Nothing.

Coin Selection

data CoinSelection i o

Represents the result of running a coin selection algorithm.

See CoinSelectionAlgorithm.

Constructors

CoinSelection

Fields

inputs :: CoinMap i

A subset of the original UTxO set that was passed to the coin selection algorithm, containing only the entries that were selected by the coin selection algorithm.

outputs :: CoinMap o

The original set of output payments passed to the coin selection algorithm, whose total value is covered by the inputs.

change :: [Coin]

A set of change values to be paid back to the originator of the payment.

sumInputs :: CoinSelection i o -> Coin

Calculate the total sum of all inputs for the given CoinSelection.

sumOutputs :: CoinSelection i o -> Coin

Calculate the total sum of all outputs for the given CoinSelection.

sumChange :: CoinSelection i o -> Coin

Calculate the total sum of all change for the given CoinSelection.

Coin Selection Algorithm

newtype CoinSelectionAlgorithm i o m e

Provides a common interface for coin selection algorithms.

The function selectCoins, when applied to the given initial UTxO set and output set, generates a CoinSelection that is capable of paying for all of the outputs, and a remaining UTxO set from which all spent values have been removed.

Each entry in the initial UTxO set refers to a unique unspent output from a previous transaction, together with its corresponding value. The algorithm will select from among the entries in this set to pay for entries in the output set, placing the selected entries in the inputs field of the resulting CoinSelection.

Each entry in the output set refers to a unique payment recipient together with the value to pay to that recipient. The outputs field of the resulting CoinSelection will be equal to this set.

The total value of the initial UTxO set must be greater than or equal to the total value of the output set, as given by the coinMapValue function.

Constructors

CoinSelectionAlgorithm

Fields

selectCoins :: CoinSelectionOptions i o e -> CoinMap i -> CoinMap o -> ExceptT (CoinSelectionError e) m (CoinSelection i o, CoinMap i)

data CoinSelectionOptions i o e

Represents a set of options to be passed to a coin selection algorithm.

Constructors

CoinSelectionOptions

Fields

maximumInputCount :: Word8 -> Word8

Calculate the maximum number of inputs allowed for a given number of outputs.

validate :: CoinSelection i o -> Either e ()

Validate the given coin selection, returning a backend-specific error.

data CoinSelectionError e

Represents the set of possible failures that can occur when attempting to produce a CoinSelection.

Constructors

ErrUtxoBalanceInsufficient Word64 Word64
The UTxO balance was insufficient to cover the total payment amount.

Records the UTxO balance, as well as the total value of the payment we tried to make.

ErrUtxoNotFragmentedEnough Word64 Word64
The UTxO was not fragmented enough to support the required number of transaction outputs.

Records the number of UTxO entries, as well as the number of the transaction outputs.

ErrUtxoFullyDepleted
Due to the particular distribution of values within the UTxO set, all available UTxO entries were depleted before all the requested transaction outputs could be paid for.

ErrMaximumInputCountExceeded Word64
The number of UTxO entries needed to cover the requested payment exceeded the upper limit specified by maximumInputCount.

Records the value of maximumInputCount.

ErrInvalidSelection e
The coin selection generated was reported to be invalid by the backend.

Records the backend-specific error that occurred while attempting to validate the selection.