Cardano.CoinSelection.Fee

Description

Provides the API of Coin Selection algorithm and Fee Calculation This module contains the implementation of adjusting coin selection for a fee. The sender pays for the fee and additional inputs are picked randomly. For more information refer to: https://iohk.io/blog/self-organisation-in-coin-selection/

Types

newtype Fee

Represents a fee to be paid on a transaction.

This type is isomorphic to Coin.

Constructors

Fee

Fields

unFee :: Word64

Fee Calculation

calculateFee :: CoinSelection i o -> Fee

Calculates the current fee associated with a given CoinSelection.

distributeFee :: Fee -> NonEmpty Coin -> NonEmpty (Fee, Coin)

Distribute the given fee over the given list of coins, so that each coin is allocated a fraction of the fee in proportion to its relative size.

Pre-condition

Every coin in the given list must be non-zero in value.

Examples

>>> distributeFee (Fee 2) [(Coin 1), (Coin 1)]
[(Fee 1, Coin 1), (Fee 1, Coin 1)]

>>> distributeFee (Fee 4) [(Coin 1), (Coin 1)]
[(Fee 2, Coin 1), (Fee 2, Coin 1)]

>>> distributeFee (Fee 7) [(Coin 1), (Coin 2), (Coin 4)]
[(Fee 1, Coin 1), (Fee 2, Coin 2), (Fee 4, Coin 4)]

>>> distributeFee (Fee 14) [(Coin 1), (Coin 2), (Coin 4)]
[(Fee 2, Coin 1), (Fee 4, Coin 2), (Fee 8, Coin 4)]

Fee Adjustment

newtype FeeEstimator i o

Provides a function capable of estimating the fee for a given coin selection.

The fee estimate can depend on the numbers of inputs, outputs, and change outputs within the coin selection, as well as their magnitudes.

Constructors

FeeEstimator

Fields

estimateFee :: CoinSelection i o -> Fee

data FeeOptions i o

Provides options for fee balancing.

Constructors

FeeOptions

Fields

feeEstimator :: FeeEstimator i o

dustThreshold :: DustThreshold

newtype ErrAdjustForFee

Constructors

ErrCannotCoverFee Word64
UTxO exhausted during fee covering We record what amount missed to cover the fee

adjustForFee :: (Buildable i, Buildable o, Ord i, MonadRandom m) => FeeOptions i o -> CoinMap i -> CoinSelection i o -> ExceptT ErrAdjustForFee m (CoinSelection i o)

Given the coin selection result from a policy run, adjust the outputs for fees, potentially returning additional inputs that we need to cover all fees.

We lose the relationship between the transaction outputs and their corresponding inputs/change outputs here. This is a decision we may wish to revisit later. For now however note that since

(a) coin selection tries to establish a particular ratio between payment outputs and change outputs (currently it aims for an average of 1:1)

(b) coin selection currently only generates a single change output per payment output, distributing the fee proportionally across all change outputs is roughly equivalent to distributing it proportionally over the payment outputs (roughly, not exactly, because the 1:1 proportion is best effort only, and may in some cases be wildly different).

Note that for (a) we don't need the ratio to be 1:1, the above reasoning will remain true for any proportion 1:n. For (b) however, if coin selection starts creating multiple outputs, and this number may vary, then losing the connection between outputs and change outputs will mean that that some outputs may pay a larger percentage of the fee (depending on how many change outputs the algorithm happened to choose).

reduceChangeOutputs :: DustThreshold -> Fee -> [Coin] -> [Coin]

Pays for the given fee by subtracting it from the given list of change outputs, so that each change output is reduced by a portion of the fee that's in proportion to its relative size.

Basic Examples

>>> reduceChangeOutputs (DustThreshold 0) (Fee 4) (Coin <$> [2, 2, 2, 2])
[Coin 1, Coin 1, Coin 1, Coin 1]

>>> reduceChangeOutputs (DustThreshold 0) (Fee 15) (Coin <$> [2, 4, 8, 16])
[Coin 1, Coin 2, Coin 4, Coin 8]

Handling Dust

Any dust outputs in the resulting list are coalesced according to the given dust threshold: (See coalesceDust.)

>>> reduceChangeOutputs (DustThreshold 1) (Fee 4) (Coin <$> [2, 2, 2, 2])
[Coin 4]

Handling Insufficient Change

If there's not enough change to pay for the fee, or if there's only just enough to pay for it exactly, this function returns the empty list:

>>> reduceChangeOutputs (DustThreshold 0) (Fee 15) (Coin <$> [10])
[]

>>> reduceChangeOutputs (DustThreshold 0) (Fee 15) (Coin <$> [1, 2, 4, 8])
[]

Dust Processing

newtype DustThreshold

Defines the maximum size of a dust coin.

Change values that are less than or equal to this threshold will not be included in transactions.

This type is isomorphic to Coin.

Constructors

DustThreshold

Fields

unDustThreshold :: Word64

coalesceDust :: DustThreshold -> NonEmpty Coin -> [Coin]

From the given list of coins, remove dust coins with a value less than or equal to the given threshold value, redistributing their total value over the coins that remain.

This function satisfies the following properties:

>>> sum coins = sum (coalesceDust threshold coins)
>>> all (/= Coin 0) (coalesceDust threshold coins)

Coin Splitting

splitCoin :: Coin -> [Coin] -> [Coin]

Splits up the given coin of value v, distributing its value over the given coin list of length n, so that each coin value is increased by an integral amount within unity of v/n, producing a new list of coin values where the overall total is preserved.

Basic Examples

When it's possible to divide a coin evenly, each coin value is increased by the same integer amount:

>>> splitCoin (Coin 40) (Coin <$> [1, 1, 1, 1])
[Coin 11, Coin 11, Coin 11, Coin 11]

>>> splitCoin (Coin 40) (Coin <$> [1, 2, 3, 4])
[Coin 11, Coin 12, Coin 13, Coin 14]

Handling Non-Uniform Increases

When it's not possible to divide a coin evenly, each integral coin value in the resulting list is always within unity of the ideal unrounded result:

>>> splitCoin (Coin 2) (Coin <$> [1, 1, 1, 1])
[Coin 1, Coin 1, Coin 2, Coin 2]

>>> splitCoin (Coin 10) (Coin <$> [1, 1, 1, 1])
[Coin 3, Coin 3, Coin 4, Coin 4]

Handling Overflow

While processing the given list, if increasing the value of any given coin c would cause its value to exceed maxBound, this function will leave coin c unchanged in the resulting list, distributing the excess value to coins that occur later in the list:

>>> splitCoin (Coin 10) (Coin <$> [unCoin maxBound, 1])
[Coin 45000000000000000, Coin 11]

>>> splitCoin (Coin 10) (Coin <$> [unCoin maxBound - 1, 1])
[Coin 44999999999999999, Coin 11]

>>> splitCoin (Coin 10) (Coin <$> [unCoin maxBound - 1, 1, 1])
[Coin 44999999999999999, Coin 6, 6]

Handling Leftover Remaining Value

If there is any remaining value left over after processing the list, a new coin is appended to the end of the list to hold the excess value:

>>> splitCoin (Coin 10) []
[Coin 10]

>>> splitCoin (Coin 10) (Coin <$> [unCoin maxBound])
[Coin 45000000000000000, Coin 10]

>>> splitCoin (Coin 10) (Coin <$> [unCoin maxBound - 1])
[Coin 44999999999999999, Coin 10]

Properties

The total value is always preserved:

>>> sum (splitCoin x ys) == x + sum ys