Cardano.CoinSelection.LargestFirst

Description

This module contains an implementation of the Largest-First coin selection algorithm.

Documentation

largestFirst :: (Ord i, Ord o, Monad m) => CoinSelectionAlgorithm i o m

An implementation of the Largest-First coin selection algorithm.

Overview

The Largest-First algorithm processes outputs in descending order of value, from largest to smallest.

For each output, it repeatedly selects the largest remaining unspent UTxO entry until the value of selected entries is greater than or equal to the value of that output.

State Maintained by the Algorithm

At all stages of processing, the algorithm maintains:

A remaining UTxO list

This is initially equal to the given initial UTxO set parameter, sorted into descending order of coin value.

The head of the list is always the remaining UTxO entry with the largest coin value.

Entries are incrementally removed from the head of the list as the algorithm proceeds, until the list is empty.

An unpaid output list

This is initially equal to the given output list parameter, sorted into descending order of coin value.

The head of the list is always the unpaid output with the largest coin value.

Entries are incrementally removed from the head of the list as the algorithm proceeds, until the list is empty.

An accumulated coin selection

This is initially empty.

Entries are incrementally added as each output is paid for, until the unpaid output list is empty.

Cardinality Rules

The algorithm requires that:

Each output from the given output list is paid for by one or more entries from the initial UTxO set.

Each entry from the initial UTxO set is used to pay for at most one output from the given output list.

(A single UTxO entry cannot be used to pay for multiple outputs.)

Order of Processing

The algorithm proceeds according to the following sequence of steps:

Step 1

Remove a single unpaid output from the head of the unpaid output list.

Step 2

Repeatedly remove UTxO entries from the head of the remaining UTxO list until the total value of entries removed is greater than or equal to the value of the removed output.

Step 3

Use the removed UTxO entries to pay for the removed output.

This is achieved by:

adding the removed UTxO entries to the inputs field of the accumulated coin selection.

adding the removed output to the outputs field of the accumulated coin selection.

Step 4

If the total value of the removed UTxO entries is greater than the value of the removed output, generate a coin whose value is equal to the exact difference, and add it to the change field of the accumulated coin selection.

Step 5

If the unpaid output list is empty, terminate here.

Otherwise, return to Step 1.

Termination

The algorithm terminates successfully if the remaining UTxO list is not depleted before the unpaid output list can be fully depleted (i.e., if all the outputs have been paid for).

The accumulated coin selection and remaining UTxO list are returned to the caller.

Failure Modes

The algorithm terminates with an error if:

The total value of the initial UTxO set (the amount of money available) is less than the total value of the output list (the amount of money required).

See: InputValueInsufficientError.

The number of entries in the initial UTxO set is smaller than the number of requested outputs.

Due to the nature of the algorithm, at least one UTxO entry is required for each output.

See: InputCountInsufficientError.

Due to the particular distribution of values within the initial UTxO set, the algorithm depletes all entries from the UTxO set before it is able to pay for all requested outputs.

See: InputsExhaustedError.

The number of UTxO entries needed to pay for the requested outputs would exceed the upper limit specified by limit.

See: InputLimitExceededError.