Internal

Description

This module hierarchy provides types and functions that are not intended to be part of the public API.

Types and functions defined herein are not guaranteed to be forwards or backwards compatible across different versions of the library.

Documentation

module Internal.Invariant

module Internal.Rounding