Cardano.CoinSelection

Description

Provides general functions and types relating to coin selection.

The CoinSelection type represents a coin selection, the basis for a transaction in a UTxO-based blockchain.

The CoinSelectionAlgorithm type provides a common interface to algorithms that generate coin selections.

For a guide to algorithms provided by this library, see the Cardano.CoinSelection.Algorithm module.

To adjust coin selections in order to pay for transaction fees, see the Cardano.CoinSelection.Fee module.

Coin Selections

data CoinSelection i o

A coin selection is the basis for a transaction.

It consists of a selection of inputs, outputs, and change.

The inputs and outputs fields are both maps of unique keys to associated Coin values, where:

Each key-value pair in the inputs map corresponds to an unspent output from a previous transaction (also known as a UTxO). The key is a unique reference to that output, and the value is the amount of unspent value associated with it.

Each key-value pair in the outputs map corresponds to a payment. The key is a unique reference to a payment recipient, and the value is the amount of money to pay to that recipient.

The change field is a set of coins to be returned to the originator of the transaction.

The CoinSelectionAlgorithm type provides a common interface for generating coin selections.

Since: 1.0.0

Constructors

CoinSelection

Fields

inputs :: CoinMap i

The set of inputs.

outputs :: CoinMap o

The set of outputs.

change :: [Coin]

The set of change.

sumInputs :: CoinSelection i o -> Coin

Calculate the total sum of all inputs for the given CoinSelection.

Since: 1.0.0

sumOutputs :: CoinSelection i o -> Coin

Calculate the total sum of all outputs for the given CoinSelection.

Since: 1.0.0

sumChange :: CoinSelection i o -> Coin

Calculate the total sum of all change for the given CoinSelection.

Since: 1.0.0

Coin Selection Algorithms

newtype CoinSelectionAlgorithm i o m

Provides a common interface for coin selection algorithms.

The function selectCoins, when applied to the given CoinSelectionParameters object (with available inputs and requested outputs), will generate a CoinSelectionResult (with remaining inputs and a coin selection).

For implementations provided by this library, see Cardano.CoinSelection.Algorithm.

Since: 1.0.0

Constructors

CoinSelectionAlgorithm

Fields

selectCoins :: CoinSelectionParameters i o -> ExceptT CoinSelectionError m (CoinSelectionResult i o)

data CoinSelectionParameters i o

The complete set of parameters required for a CoinSelectionAlgorithm.

The inputsAvailable and outputsRequested fields are both maps of unique keys to associated Coin values, where:

Each key-value pair in the inputsAvailable map corresponds to an unspent output from a previous transaction that is available for selection as an input by the coin selection algorithm. The key is a unique reference to that output, and the value is the amount of unspent value associated with it.

Each key-value pair in the outputsRequested map corresponds to a payment whose value is to be paid for by the coin selection algorithm. The key is a unique reference to a payment recipient, and the value is the amount of money to pay to that recipient.

A coin selection algorithm will select a subset of inputs from inputsAvailable in order to pay for all the outputs in outputsRequested, where:

Inputs selected by the algorithm are included in the inputs set of the generated CoinSelection.

Inputs not selected by the algorithm are included in the inputsRemaining set of the CoinSelectionResult.

The number of inputs that can selected is limited by limit.

The total value of inputsAvailable must be greater than or equal to the total value of outputsRequested, as given by the coinMapValue function.

Since: 1.0.0

Constructors

CoinSelectionParameters

Fields

inputsAvailable :: CoinMap i

The set of inputs available for selection.

outputsRequested :: CoinMap o

The set of outputs requested for payment.

limit :: CoinSelectionLimit

A limit on the number of inputs that can be selected.

data CoinSelectionResult i o

Represents the result of running a coin selection algorithm.

See CoinSelectionAlgorithm.

Since: 1.0.0

Constructors

CoinSelectionResult

Fields

coinSelection :: CoinSelection i o

The generated coin selection.

inputsRemaining :: CoinMap i

The set of inputs that were not selected.

newtype CoinSelectionLimit

Defines an inclusive upper bound on the number of inputs that a CoinSelectionAlgorithm is allowed to select.

Since: 1.0.0

Constructors

CoinSelectionLimit

Fields

calculateLimit :: Word16 -> Word16

Calculate the maximum number of inputs allowed for a given number of outputs.

Coins

data Coin

Represents a non-negative integral amount of currency.

Use coinFromNatural to create a coin from a natural number.

Use coinToNatural to convert a coin into a natural number.

Since: 1.0.0

coinFromNatural :: Natural -> Coin

Creates a coin from a natural number.

Since: 1.0.0

coinToNatural :: Coin -> Natural

Converts the given coin into a natural number.

Since: 1.0.0

Coin Maps

newtype CoinMap a

A mapping from unique keys to associated Coin values.

A CoinMap can be used to represent:

a UTxO set, where each key within the map refers to an unspent output from a previous transaction.

a set of inputs to a CoinSelection, where each input is an entry selected from a UTxO set by a CoinSelectionAlgorithm.

a set of outputs for a CoinSelection, where each key within the map refers to the address of a payment recipient.

A CoinMap can be constructed with the coinMapFromList function.

The total value of a CoinMap is given by the coinMapValue function.

Since: 1.0.0

Constructors

CoinMap

Fields

unCoinMap :: Map a Coin

data CoinMapEntry a

An entry for a CoinMap.

Since: 1.0.0

Constructors

CoinMapEntry

Fields

entryKey :: a

The unique key associated with this entry.

entryValue :: Coin

The coin value associated with this entry.

coinMapFromList :: Ord a => [CoinMapEntry a] -> CoinMap a

Constructs a CoinMap from a list of entries.

See CoinMapEntry.

Since: 1.0.0

coinMapToList :: CoinMap a -> [CoinMapEntry a]

Converts a CoinMap to a list of entries.

See CoinMapEntry.

Since: 1.0.0

coinMapValue :: CoinMap a -> Coin

Calculates the total coin value associated with a CoinMap.

Since: 1.0.0

Coin Selection Errors

data CoinSelectionError

Represents the set of possible failures that can occur when attempting to produce a CoinSelection with a CoinSelectionAlgorithm.

See selectCoins.

Since: 1.0.0

Constructors

InputValueInsufficient InputValueInsufficientError

InputCountInsufficient InputCountInsufficientError

InputLimitExceeded InputLimitExceededError

InputsExhausted InputsExhaustedError

data InputValueInsufficientError

Indicates that the total value of inputsAvailable is less than the total value of outputsRequested, making it impossible to cover all payments, regardless of which algorithm is chosen.

Since: 1.0.0

Constructors

InputValueInsufficientError

Fields

inputValueAvailable :: Coin

The total value of inputsAvailable.

inputValueRequired :: Coin

The total value of outputsRequested.

data InputCountInsufficientError

Indicates that the total count of entries in inputsAvailable is fewer than required by the algorithm. The number required depends on the particular algorithm implementation.

Since: 1.0.0

Constructors

InputCountInsufficientError

Fields

inputCountAvailable :: Natural

The number of entries in inputsAvailable.

inputCountRequired :: Natural

The number of entries required.

newtype InputLimitExceededError

Indicates that the coin selection algorithm is unable to cover the total value of outputsRequested without exceeding the maximum number of inputs defined by limit.

See calculateLimit.

Since: 1.0.0

Constructors

InputLimitExceededError

Fields

calculatedInputLimit :: Word16

data InputsExhaustedError

Indicates that all available entries in inputsAvailable were depleted before all the payments in outputsRequested could be paid for.

This condition can occur even if the total value of inputsAvailable is greater than or equal to the total value of outputsRequested, due to differences in the way that algorithms select inputs.

Since: 1.0.0

Constructors

InputsExhaustedError