Cardano.CoinSelection.Fee

Description

Provides functionality for adjusting coin selections in order to pay for transaction fees.

Fundamental Types

newtype Fee

Represents a non-negative fee to be paid on a transaction.

Constructors

Fee

Fields

unFee :: Coin

newtype FeeEstimator i o

Provides a function capable of estimating the fee for a given coin selection.

The fee estimate can depend on the numbers of inputs, outputs, and change outputs within the coin selection, as well as their magnitudes.

Constructors

FeeEstimator

Fields

estimateFee :: CoinSelection i o -> Fee

newtype DustThreshold

Defines the maximum size of a dust coin.

Change values that are less than or equal to this threshold will not be included in coin selections produced by the adjustForFee function.

Constructors

DustThreshold

Fields

unDustThreshold :: Coin

Fee Adjustment

adjustForFee :: (Ord i, Show i, Show o, MonadRandom m) => FeeOptions i o -> CoinMap i -> CoinSelection i o -> ExceptT FeeAdjustmentError m (CoinSelection i o)

Given the coin selection result from a policy run, adjust the outputs for fees, potentially returning additional inputs that we need to cover all fees.

We lose the relationship between the transaction outputs and their corresponding inputs/change outputs here. This is a decision we may wish to revisit later. For now however note that since

(a) coin selection tries to establish a particular ratio between payment outputs and change outputs (currently it aims for an average of 1:1)

(b) coin selection currently only generates a single change output per payment output, distributing the fee proportionally across all change outputs is roughly equivalent to distributing it proportionally over the payment outputs (roughly, not exactly, because the 1:1 proportion is best effort only, and may in some cases be wildly different).

Note that for (a) we don't need the ratio to be 1:1, the above reasoning will remain true for any proportion 1:n. For (b) however, if coin selection starts creating multiple outputs, and this number may vary, then losing the connection between outputs and change outputs will mean that that some outputs may pay a larger percentage of the fee (depending on how many change outputs the algorithm happened to choose).

data FeeOptions i o

Provides options for fee adjustment.

Constructors

FeeOptions

Fields

feeEstimator :: FeeEstimator i o

dustThreshold :: DustThreshold